[refactor] Rename package; net.honux.chess -> chess.pieces
