+++ src/main/java/chess/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package net.honux.chess;
+package chess.pieces;
 
 public class Pawn {
     private final Color color;
